services/user_service/requirements.txt
alembic>=1.7.1
asyncpg>=0.25.0
emails>=0.6
fastapi>=0.68.0
passlib[bcrypt]>=1.7.4
psycopg2-binary>=2.9.1
pydantic>=1.8.2
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
sqlalchemy>=1.4.23
sqlalchemy-utils>=0.38.2
uvicorn>=0.15.0
pytest 
pytest-asyncio 
pytest-cov 
httpx
PyJWT>=2.4.0
bcrypt>=3.2.0
email-validator>=1.1.3